+++ src/main/java/com/redhat/syseng/openshift/service/broker/model/amp/Metric.java
@@ -24,11 +24,11 @@ import javax.xml.bind.annotation.XmlType;
  *   &lt;complexContent>
  *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
  *       &lt;sequence>
- *         &lt;element name="id" type="{http://www.w3.org/2001/XMLSchema}byte"/>
+ *         &lt;element name="id" type="{http://www.w3.org/2001/XMLSchema}int"/>
  *         &lt;element name="name" type="{http://www.w3.org/2001/XMLSchema}string"/>
  *         &lt;element name="system_name" type="{http://www.w3.org/2001/XMLSchema}string"/>
  *         &lt;element name="friendly_name" type="{http://www.w3.org/2001/XMLSchema}string"/>
- *         &lt;element name="service_id" type="{http://www.w3.org/2001/XMLSchema}byte"/>
+ *         &lt;element name="service_id" type="{http://www.w3.org/2001/XMLSchema}int"/>
  *         &lt;element name="unit" type="{http://www.w3.org/2001/XMLSchema}string"/>
  *       &lt;/sequence>
  *     &lt;/restriction>
@@ -49,7 +49,7 @@ import javax.xml.bind.annotation.XmlType;
 })
 public class Metric {
 
-    protected byte id;
+    protected int id;
     @XmlElement(required = true)
     protected String name;
     @XmlElement(name = "system_name", required = true)
@@ -57,7 +57,7 @@ public class Metric {
     @XmlElement(name = "friendly_name", required = true)
     protected String friendlyName;
     @XmlElement(name = "service_id")
-    protected byte serviceId;
+    protected int serviceId;
     @XmlElement(required = true)
     protected String unit;
 
@@ -65,7 +65,7 @@ public class Metric {
      * Gets the value of the id property.
      * 
      */
-    public byte getId() {
+    public int getId() {
         return id;
     }
 
@@ -73,7 +73,7 @@ public class Metric {
      * Sets the value of the id property.
      * 
      */
-    public void setId(byte value) {
+    public void setId(int value) {
         this.id = value;
     }
 
@@ -153,7 +153,7 @@ public class Metric {
      * Gets the value of the serviceId property.
      * 
      */
-    public byte getServiceId() {
+    public int getServiceId() {
         return serviceId;
     }
 
@@ -161,7 +161,7 @@ public class Metric {
      * Sets the value of the serviceId property.
      * 
      */
-    public void setServiceId(byte value) {
+    public void setServiceId(int value) {
         this.serviceId = value;
     }
