+++ src/main/java/com/redhat/syseng/openshift/service/broker/model/amp/Plan.java
@@ -26,16 +26,16 @@ import javax.xml.bind.annotation.XmlType;
  *   &lt;complexContent>
  *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType">
  *       &lt;sequence>
- *         &lt;element name="id" type="{http://www.w3.org/2001/XMLSchema}byte"/>
+ *         &lt;element name="id" type="{http://www.w3.org/2001/XMLSchema}int"/>
  *         &lt;element name="name" type="{http://www.w3.org/2001/XMLSchema}string"/>
  *         &lt;element name="type" type="{http://www.w3.org/2001/XMLSchema}string"/>
  *         &lt;element name="state" type="{http://www.w3.org/2001/XMLSchema}string"/>
- *         &lt;element name="service_id" type="{http://www.w3.org/2001/XMLSchema}byte"/>
+ *         &lt;element name="service_id" type="{http://www.w3.org/2001/XMLSchema}int"/>
  *         &lt;element name="end_user_required" type="{http://www.w3.org/2001/XMLSchema}string"/>
  *         &lt;element name="setup_fee" type="{http://www.w3.org/2001/XMLSchema}float"/>
  *         &lt;element name="cost_per_month" type="{http://www.w3.org/2001/XMLSchema}float"/>
- *         &lt;element name="trial_period_days" type="{http://www.w3.org/2001/XMLSchema}byte"/>
- *         &lt;element name="cancellation_period" type="{http://www.w3.org/2001/XMLSchema}byte"/>
+ *         &lt;element name="trial_period_days" type="{http://www.w3.org/2001/XMLSchema}int"/>
+ *         &lt;element name="cancellation_period" type="{http://www.w3.org/2001/XMLSchema}int"/>
  *       &lt;/sequence>
  *       &lt;attribute name="custom" type="{http://www.w3.org/2001/XMLSchema}string" />
  *       &lt;attribute name="default" type="{http://www.w3.org/2001/XMLSchema}string" />
@@ -62,7 +62,7 @@ import javax.xml.bind.annotation.XmlType;
 @XmlRootElement(name = "plan")
 public class Plan {
 
-    protected byte id;
+    protected int id;
     @XmlElement(required = true)
     protected String name;
     @XmlElement(required = true)
@@ -70,7 +70,7 @@ public class Plan {
     @XmlElement(required = true)
     protected String state;
     @XmlElement(name = "service_id")
-    protected byte serviceId;
+    protected int serviceId;
     @XmlElement(name = "end_user_required", required = true)
     protected String endUserRequired;
     @XmlElement(name = "setup_fee")
@@ -78,9 +78,9 @@ public class Plan {
     @XmlElement(name = "cost_per_month")
     protected float costPerMonth;
     @XmlElement(name = "trial_period_days")
-    protected byte trialPeriodDays;
+    protected int trialPeriodDays;
     @XmlElement(name = "cancellation_period")
-    protected byte cancellationPeriod;
+    protected int cancellationPeriod;
     @XmlAttribute(name = "custom")
     protected String custom;
     @XmlAttribute(name = "default")
@@ -90,7 +90,7 @@ public class Plan {
      * Gets the value of the id property.
      * 
      */
-    public byte getId() {
+    public int getId() {
         return id;
     }
 
@@ -98,7 +98,7 @@ public class Plan {
      * Sets the value of the id property.
      * 
      */
-    public void setId(byte value) {
+    public void setId(int value) {
         this.id = value;
     }
 
@@ -178,7 +178,7 @@ public class Plan {
      * Gets the value of the serviceId property.
      * 
      */
-    public byte getServiceId() {
+    public int getServiceId() {
         return serviceId;
     }
 
@@ -186,7 +186,7 @@ public class Plan {
      * Sets the value of the serviceId property.
      * 
      */
-    public void setServiceId(byte value) {
+    public void setServiceId(int value) {
         this.serviceId = value;
     }
 
@@ -250,7 +250,7 @@ public class Plan {
      * Gets the value of the trialPeriodDays property.
      * 
      */
-    public byte getTrialPeriodDays() {
+    public int getTrialPeriodDays() {
         return trialPeriodDays;
     }
 
@@ -258,7 +258,7 @@ public class Plan {
      * Sets the value of the trialPeriodDays property.
      * 
      */
-    public void setTrialPeriodDays(byte value) {
+    public void setTrialPeriodDays(int value) {
         this.trialPeriodDays = value;
     }
 
@@ -266,7 +266,7 @@ public class Plan {
      * Gets the value of the cancellationPeriod property.
      * 
      */
-    public byte getCancellationPeriod() {
+    public int getCancellationPeriod() {
         return cancellationPeriod;
     }
 
@@ -274,7 +274,7 @@ public class Plan {
      * Sets the value of the cancellationPeriod property.
      * 
      */
-    public void setCancellationPeriod(byte value) {
+    public void setCancellationPeriod(int value) {
         this.cancellationPeriod = value;
     }
